build: rename index.js to vuex-cache.js and change package entries (thx bili)
